refactor(subagents): clarify runner boundaries
